[ARA] Instead of generating a tempo map by walking each beat, walk the tempo entries and handle curves and edge cases.
To handle curves, keep a running index into the array of tempo sections and use that data to create entries between tempo curves. Don't include curve data if the tempo setting is at the end of the requested range. There's one slight oddity here - if the curve value is -1, right now the code may give the same tempo value twice at the end. We could use a curve of -1 as a cue to skip that tempo setting, but wasn't sure if that behavior would be correct in all cases.

To allow plug-ins to compute tempo at the end the range or beyond the last tempo setting we can either include one extra setting, or extrapolate a new entry one minute ahead.