[ARA] Adding new reader classes for key signatures and chords using the pitch sequence and chord track.
To do this we had to add some new utility functions to help map the classes and data in tracktion_Musicality to their ARA representations. I also found it useful to make the PitchSequence::indexOfPitchAt function public, mirroring a similar function in TempoSequence. I've verified this using a tracktion project edit with chord and key signature changes that I loaded into the PitchAndTimeDemo example, and things seem to be working as expected.